Separated the code into packages to allow for easier access, usage, and testing
