+++ tests/testContactTracer.py
@@ -1,0 +1,9 @@
+import unittest
+
+class TestContactTracer(unittest.TestCase):
+    def test_something(self):
+        self.assertEqual(True, False)  # add assertion here
+
+
+if __name__ == '__main__':
+    unittest.main()
